I have added authentication for user and as well as journal entires. And also encrypted it. So only people with username and Password can now access it. I have just added spring authentication so basically basic http authentication only. For the next step I will be adding role based authentication

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -6,6 +6,7 @@ import net.engineeringdigest.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -18,11 +19,18 @@ public class JournalEntryService {
     @Autowired
     private UserService userService;
 
+    @Transactional //if we give transactional property like this it will implement the entire code in one go
+            //or none of the code. so it is like acid properties
     public void saveEntry(JournalEntry entry, String userName) {
-        User user = userService.findByUsername(userName);
-        JournalEntry savedEntry = repository.save(entry); //ami ai2 save hua journal entry tu lolu
-        user.getJournalEntries().add(savedEntry);  //xe2 user r in memory t save korilu
-        userService.saveEntry(user);  // and xe2 db t save krlu
+        try{
+            User user = userService.findByUsername(userName);
+            JournalEntry savedEntry = repository.save(entry); //ami ai2 save hua journal entry tu lolu
+            user.getJournalEntries().add(savedEntry);//xe2 user r in memory t save korilu
+            userService.saveUser(user);
+        }catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException(e);
+        }
     }
 
     public void saveEntry(JournalEntry entry) {
@@ -37,11 +45,22 @@ public class JournalEntryService {
         return repository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName) {
-        User user = userService.findByUsername(userName);
-        user.getJournalEntries().removeIf(entry -> entry.getId().equals(id));//user tur pora delete kori save krilu karon
-        //user t thoka journal entry nije nije delete nhi
-        repository.deleteById(id);
-    }
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try{
+            User user = userService.findByUsername(userName);
+            removed = user.getJournalEntries().removeIf(entry -> entry.getId().equals(id));//user tur pora delete kori save krilu karon
+            //user t thoka journal entry nije nije delete nhi
+            if(removed) {
+                userService.saveUser(user);
+                repository.deleteById(id);
+            }
 
+        }catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting journal entry "+ e);
+        }
+       return removed;
+    }
 }
